Extracted FilePathRange and Pool into separate files.
